Lab_4/requirements.txt
alembic
annotated-types
anyio
click
fastapi
h11
h2
hpack
httpcore
idna
jinja2
jwt
passlib
pydantic
pydantic_core
PyJWT
python-jose==3.2.0
python-multipart
PyYAML
setuptools==75.8.0
sniffio
starlette
sqlalchemy
psycopg2-binary
uvicorn
wheel==0.45.1
motor